backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
azure-servicebus==7.11.0
azure-cosmos==4.4.0
azure-storage-blob==12.18.0
openai==1.3.0
anthropic==0.7.1
tenacity==8.2.3
python-json-logger==2.0.7
